qcom-common: Sync source tree with caf/jb_1.9.6

+++ rootdir/etc/init.qcom.wifi.sh
@@ -60,7 +60,7 @@ trigger_wcnss()
             # caldata is available at userspace.
             if [ -e /data/misc/wifi/WCNSS_qcom_wlan_cal.bin ]; then
                 calparm=`ls /sys/module/wcnsscore/parameters/has_calibrated_data`
-                if [ -e $calparm ]; then
+                if [ -e $calparm ] && [ ! -e /data/misc/wifi/WCN_FACTORY ]; then
                     echo 1 > $calparm
                 fi
             fi
@@ -89,196 +89,22 @@ trigger_wcnss()
 
 
 case "$target" in
-    msm8974* | msm8226* | msm8610*)
-
-# Check whether device is plugged on the HSIC bus
-# Currently HSIC bus will be the first index
-
-    if [ -e /sys/bus/platform/drivers/msm_hsic_host ]; then
-       if [ ! -L /sys/bus/usb/devices/1-1 ]; then
-           echo msm_hsic_host > /sys/bus/platform/drivers/msm_hsic_host/unbind
-       else
-           echo auto > /sys/bus/usb/devices/1-1/power/control
-       fi
-
-       chmod 0222 /sys/bus/platform/drivers/msm_hsic_host/bind
-       chmod 0222 /sys/bus/platform/drivers/msm_hsic_host/unbind
-    fi
-
-    wlanchip=""
-
-# force ar6004 is ar6004_wlan.conf existed.
-    if [ -f /system/etc/firmware/ath6k/AR6004/ar6004_wlan.conf ]; then
-        wlanchip=`cat /system/etc/firmware/ath6k/AR6004/ar6004_wlan.conf`
-    fi
-
-# auto detect ar6004-sdio card
-# for ar6004-sdio card, the vendor id and device id is as the following
-# vendor id  device id
-#    0x0271     0x0400
-#    0x0271     0x0401
-    if [ "$wlanchip" == "" ]; then
-        sdio_vendors=`echo \`cat /sys/bus/mmc/devices/*/*/vendor\``
-        sdio_devices=`echo \`cat /sys/bus/mmc/devices/*/*/device\``
-        ven_idx=0
-
-        for vendor in $sdio_vendors; do
-            case "$vendor" in
-            "0x0271")
-                dev_idx=0
-                for device in $sdio_devices; do
-                    if [ $ven_idx -eq $dev_idx ]; then
-                        case "$device" in
-                        "0x0400" | "0x0401")
-                            wlanchip="AR6004-SDIO"
-                            ;;
-                        *)
-                            ;;
-                        esac
-                    fi
-                    dev_idx=$(( $dev_idx + 1))
-                done
-                ;;
-            *)
-                ;;
-            esac
-            ven_idx=$(( $ven_idx + 1))
-        done
-    # auto detect ar6004-sdio card end
-    fi
-
-# for ar6004-usb card, the vendor id and device id is as the following
-# vendor id  product id
-#    0x0cf3     0x9374
-#    0x0cf3     0x9372
-    if [ "$wlanchip" == "" ]; then
-        usb_vendors=`echo \`cat /sys/bus/usb/devices/*/*/idVendor\``
-        usb_products=`echo \`cat /sys/bus/usb/devices/*/*/idProduct\``
-        ven_idx=0
-
-        for vendor in $usb_vendors; do
-            case "$vendor" in
-            "0cf3")
-                dev_idx=0
-                for product in $usb_products; do
-                    if [ $ven_idx -eq $dev_idx ]; then
-                        case "$product" in
-                        "9374" | "9372")
-                            wlanchip="AR6004-USB"
-                            ;;
-                        *)
-                            ;;
-                        esac
-                    fi
-                    dev_idx=$(( $dev_idx + 1))
-                done
-                ;;
-            *)
-                ;;
-            esac
-            ven_idx=$(( $ven_idx + 1))
-        done
-    # auto detect ar6004-usb card end
-    fi
-
-      echo "The WLAN Chip ID is $wlanchip"
-      case "$wlanchip" in
-      "AR6004-USB")
-      echo msm_hsic_host > /sys/bus/platform/drivers/msm_hsic_host/unbind
-      setprop wlan.driver.ath 2
-      setprop qcom.bluetooth.soc ath3k
-      rm  /system/lib/modules/wlan.ko
-      ln -s /system/lib/modules/ath6kl-3.5/ath6kl_usb.ko \
-		/system/lib/modules/wlan.ko
-      rm /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
-      rm /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin_usb \
-		/system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin_usb \
-		/system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-      rm /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin_usb \
-                /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
-
-      # Use different wpa_supplicant.conf template between wcn driver
-      # and ath6kl driver
-      rm /system/etc/wifi/wpa_supplicant.conf
-      ln -s /system/etc/wifi/wpa_supplicant_ath6kl.conf \
-                /system/etc/wifi/wpa_supplicant.conf
-      ;;
-
-      "AR6004-SDIO")
-      setprop wlan.driver.ath 2
-      setprop qcom.bluetooth.soc ath3k
-      # Chown polling nodes as needed from UI running on system server
-      chmod 0664 /sys/devices/msm_sdcc.1/polling
-      chmod 0664 /sys/devices/msm_sdcc.2/polling
-      chmod 0664 /sys/devices/msm_sdcc.3/polling
-      chmod 0664 /sys/devices/msm_sdcc.4/polling
-
-      chown system system /sys/devices/msm_sdcc.1/polling
-      chown system system /sys/devices/msm_sdcc.2/polling
-      chown system system /sys/devices/msm_sdcc.3/polling
-      chown system system /sys/devices/msm_sdcc.4/polling
-
-      rm  /system/lib/modules/wlan.ko
-      ln -s /system/lib/modules/ath6kl-3.5/ath6kl_sdio.ko \
-		/system/lib/modules/wlan.ko
-      rm /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
-      rm /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin_sdio \
-		/system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin_sdio \
-		/system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-      ln -s /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin_sdio \
-                /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
-
-      # Use different wpa_supplicant.conf template between wcn driver
-      # and ath6kl driver
-      rm /system/etc/wifi/wpa_supplicant.conf
-      ln -s /system/etc/wifi/wpa_supplicant_ath6kl.conf \
-                /system/etc/wifi/wpa_supplicant.conf
-      ;;
-
-      *)
-      echo "*** WI-FI chip ID is not specified in /persist/wlan_chip_id **"
-      echo "*** Use the default WCN driver.                             **"
-      setprop wlan.driver.ath 0
-      rm  /system/lib/modules/wlan.ko
-      ln -s /system/lib/modules/pronto/pronto_wlan.ko \
-		/system/lib/modules/wlan.ko
-      # Populate the writable driver configuration file
-      if [ ! -e /data/misc/wifi/WCNSS_qcom_cfg.ini ]; then
-          cp /system/etc/wifi/WCNSS_qcom_cfg.ini \
-		/data/misc/wifi/WCNSS_qcom_cfg.ini
-          chown system:wifi /data/misc/wifi/WCNSS_qcom_cfg.ini
-          chmod 660 /data/misc/wifi/WCNSS_qcom_cfg.ini
-      fi
+    msm8974*)
+      # link pronto modules
+      rm /system/lib/modules/wlan.ko
+      rm /system/lib/modules/cfg80211.ko
+      ln -s /system/lib/modules/pronto/pronto_wlan.ko /system/lib/modules/wlan.ko
+      ln -s /system/lib/modules/pronto/cfg80211.ko /system/lib/modules/cfg80211.ko
 
       # The property below is used in Qcom SDK for softap to determine
       # the wifi driver config file
       setprop wlan.driver.config /data/misc/wifi/WCNSS_qcom_cfg.ini
 
-      # Use different wpa_supplicant.conf template between wcn driver
-      # and ath6kl driver
-      rm /system/etc/wifi/wpa_supplicant.conf
-      ln -s /system/etc/wifi/wpa_supplicant_wcn.conf \
-                /system/etc/wifi/wpa_supplicant.conf
-
       # Trigger WCNSS platform driver
       trigger_wcnss &
       ;;
-      esac
-      ;;
 
     msm8960*)
-
-      # Move cfg80211.ko to prima directory, the default cfg80211.ko is
-      # for wcnss solution
-      if [ ! -L /system/lib/modules/cfg80211.ko ]; then
-          mv /system/lib/modules/cfg80211.ko /system/lib/modules/prima/
-      fi
-
       wlanchip=""
 
       if [ -f /system/etc/firmware/ath6k/AR6004/ar6004_wlan.conf ]; then
@@ -291,39 +117,28 @@ case "$target" in
           # vendor id  product id
           #    0x0cf3     0x9374
           #    0x0cf3     0x9372
-          usb_vendors=`echo \`cat /sys/bus/usb/devices/*/*/idVendor\``
-          usb_products=`echo \`cat /sys/bus/usb/devices/*/*/idProduct\``
-          ven_idx=0
-
-          for vendor in $usb_vendors; do
-              case "$vendor" in
-              "0cf3")
-                  dev_idx=0
-                  for product in $usb_products; do
-                      if [ $ven_idx -eq $dev_idx ]; then
-                          case "$product" in
-                          "9374" | "9372")
+          device_ids=`ls /sys/bus/usb/devices/`
+          for id in $device_ids; do
+              if [ -f /sys/bus/usb/devices/$id/idVendor ]; then
+                  vendor=`cat /sys/bus/usb/devices/$id/idVendor`
+                  if [ $vendor = "0cf3" ]; then
+                      if [ -f /sys/bus/usb/devices/$id/idProduct ]; then
+                          product=`cat /sys/bus/usb/devices/$id/idProduct`
+                          if [ $product = "9374" ] || [ $product = "9372" ]; then
+                              echo "auto" > /sys/bus/usb/devices/$id/power/control
                               wlanchip="AR6004-USB"
-                              ;;
-                          *)
-                              ;;
-                          esac
+                              break
+                          fi
                       fi
-                      dev_idx=$(( $dev_idx + 1))
-                  done
-                  ;;
-              *)
-                  ;;
-              esac
-              ven_idx=$(( $ven_idx + 1))
+                  fi
+              fi
           done
           # auto detect ar6004-usb card end
       fi
 
       if [ "$wlanchip" == "" ]; then
           # auto detect ar6004-sdio card
-          # for ar6004-sdio card, the vendor id and device id is
-          # as the following
+          # for ar6004-sdio card, the vendor id and device id is as the following
           # vendor id  device id
           #    0x0271     0x0400
           #    0x0271     0x0401
@@ -357,71 +172,61 @@ case "$target" in
       fi
 
       echo "The WLAN Chip ID is $wlanchip"
+      setprop wlan.driver.ath.wlanchip $wlanchip
       case "$wlanchip" in
       "AR6004-USB")
         setprop wlan.driver.ath 2
         rm  /system/lib/modules/wlan.ko
-        rm  /system/lib/modules/cfg80211.ko
         ln -s /system/lib/modules/ath6kl-3.5/ath6kl_usb.ko \
 		/system/lib/modules/wlan.ko
-        ln -s /system/lib/modules/ath6kl-3.5/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
         rm /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
         rm /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
         ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin_usb \
 		/system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
         ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin_usb \
 		/system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-
-        # Use different wpa_supplicant.conf template between wcn driver
-        # and ath6kl driver
-        rm /system/etc/wifi/wpa_supplicant.conf
-        ln -s /system/etc/wifi/wpa_supplicant_ath6kl.conf \
-                /system/etc/wifi/wpa_supplicant.conf
+        rm /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
+        ln -s /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin_usb \
+		/system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
         ;;
       "AR6004-SDIO")
         setprop wlan.driver.ath 2
         setprop qcom.bluetooth.soc ath3k
         btsoc="ath3k"
         rm  /system/lib/modules/wlan.ko
-        rm  /system/lib/modules/cfg80211.ko
         ln -s /system/lib/modules/ath6kl-3.5/ath6kl_sdio.ko \
 		/system/lib/modules/wlan.ko
-        ln -s /system/lib/modules/ath6kl-3.5/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
         rm /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
         rm /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
         ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin_sdio \
 		/system/etc/firmware/ath6k/AR6004/hw1.3/fw.ram.bin
         ln -s /system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin_sdio \
 		/system/etc/firmware/ath6k/AR6004/hw1.3/bdata.bin
-
-        # Use different wpa_supplicant.conf template between wcn driver
-        # and ath6kl driver
-        rm /system/etc/wifi/wpa_supplicant.conf
-        ln -s /system/etc/wifi/wpa_supplicant_ath6kl.conf \
-                  /system/etc/wifi/wpa_supplicant.conf
+        rm /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
+        ln -s /system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin_sdio \
+		/system/etc/firmware/ath6k/AR6004/hw3.0/bdata.bin
         ;;
       *)
         echo "*** WI-FI chip ID is not specified in /persist/wlan_chip_id **"
         echo "*** Use the default WCN driver.                             **"
-        setprop wlan.driver.ath 0
+        setprop wlan.driver.ath 0 
         rm  /system/lib/modules/wlan.ko
-        rm  /system/lib/modules/cfg80211.ko
-        ln -s /system/lib/modules/prima/prima_wlan.ko \
-		/system/lib/modules/wlan.ko
-        ln -s /system/lib/modules/prima/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
+        ln -s /system/lib/modules/prima/prima_wlan.ko /system/lib/modules/wlan.ko
+        ln -s /system/lib/modules/prima/cfg80211.ko /system/lib/modules/cfg80211.ko
+        # Populate the writable driver configuration file
+        if [ ! -e /data/misc/wifi/WCNSS_qcom_cfg.ini ]; then
+            if [ -f /persist/WCNSS_qcom_cfg.ini ]; then
+                cp /persist/WCNSS_qcom_cfg.ini /data/misc/wifi/WCNSS_qcom_cfg.ini
+            else
+                cp /system/etc/wifi/WCNSS_qcom_cfg.ini /data/misc/wifi/WCNSS_qcom_cfg.ini
+            fi
+            chown system:wifi /data/misc/wifi/WCNSS_qcom_cfg.ini
+            chmod 660 /data/misc/wifi/WCNSS_qcom_cfg.ini
+        fi
 
         # The property below is used in Qcom SDK for softap to determine
         # the wifi driver config file
         setprop wlan.driver.config /data/misc/wifi/WCNSS_qcom_cfg.ini
-
-        # Use different wpa_supplicant.conf template between wcn driver
-        # and ath6kl driver
-        rm /system/etc/wifi/wpa_supplicant.conf
-        ln -s /system/etc/wifi/wpa_supplicant_wcn.conf \
-                  /system/etc/wifi/wpa_supplicant.conf
 
         # Trigger WCNSS platform driver
         trigger_wcnss &
@@ -430,12 +235,6 @@ case "$target" in
       ;;
 
     msm7627a*)
-
-        # The default cfg80211 module is for volans
-        if [ ! -L /system/lib/modules/cfg80211.ko ]; then
-            mv /system/lib/modules/cfg80211.ko /system/lib/modules/volans/
-        fi
-
         wlanchip=`cat /persist/wlan_chip_id`
         echo "The WLAN Chip ID is $wlanchip"
         case "$wlanchip" in
@@ -443,32 +242,25 @@ case "$target" in
              setprop wlan.driver.ath 1
              rm  /system/lib/modules/wlan.ko
              rm  /system/lib/modules/cfg80211.ko
-             ln -s /system/lib/modules/ath6kl/ath6kl_sdio.ko \
-		/system/lib/modules/wlan.ko
-             ln -s /system/lib/modules/ath6kl/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
+             ln -s /system/lib/modules/ath6kl/ath6kl_sdio.ko /system/lib/modules/wlan.ko
+             ln -s /system/lib/modules/ath6kl/cfg80211.ko /system/lib/modules/cfg80211.ko
              ;;
             "WCN1314")
              setprop wlan.driver.ath 0
              rm  /system/lib/modules/wlan.ko
              rm  /system/lib/modules/cfg80211.ko
-             ln -s /system/lib/modules/volans/WCN1314_rf.ko \
-		/system/lib/modules/wlan.ko
-             ln -s /system/lib/modules/volans/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
+             ln -s /system/lib/modules/volans/WCN1314_rf.ko /system/lib/modules/wlan.ko
+             ln -s /system/lib/modules/volans/cfg80211.ko /system/lib/modules/cfg80211.ko
              ;;
             *)
              setprop wlan.driver.ath 1
              rm  /system/lib/modules/wlan.ko
              rm  /system/lib/modules/cfg80211.ko
-             ln -s /system/lib/modules/ath6kl/ath6kl_sdio.ko \
-		/system/lib/modules/wlan.ko
-             ln -s /system/lib/modules/ath6kl/cfg80211.ko \
-		/system/lib/modules/cfg80211.ko
-             echo "************************************************************"
-             echo "*** Error:WI-FI chip ID is not specified in"
-             echo "/persist/wlan_chip_id"
-             echo "*******    WI-FI may not work    ***************************"
+             ln -s /system/lib/modules/ath6kl/ath6kl_sdio.ko /system/lib/modules/wlan.ko
+             ln -s /system/lib/modules/ath6kl/cfg80211.ko /system/lib/modules/cfg80211.ko
+             echo "********************************************************************"
+             echo "*** Error:WI-FI chip ID is not specified in /persist/wlan_chip_id **"
+             echo "*******    WI-FI may not work    ***********************************"
              ;;
         esac
     ;;
